+++ setupTests.js
@@ -21,7 +21,7 @@ expect.extend({
 
 ModalDialog.setAppElement(document.body);
 
-URL.createObjectURL = object => `blob:${object}`;
+URL.createObjectURL = object => object;
 URL.revokeObjectURL = () => {};
 
 window.requestAnimationFrame = callback => callback();